+++ src/canvas/word.jsx
@@ -57,7 +57,7 @@ function Cloud({ count = 4, radius = 20 }) {
 
 export default function Words() {
   return (
-    <Canvas dpr={[1, 2]} camera={{ position: [0, 0, 35], fov: 90 }} scale={[3, 3, 3]}>
+    <Canvas dpr={[1, 2]} camera={{ position: [0, 0, 35], fov: 90 }} scale={[3, 3, 3]} >
       <fog attach="fog" args={['#202025', 0, 80]} />
       <Cloud count={8} radius={20} />
       <TrackballControls />
