+++ tailwind.config.js
@@ -7,7 +7,7 @@ module.exports = {
       colors: {
         primary: "#050816",
         secondary: "#aaa6c3",
-        tertiary: "#151030",
+        tertiary: "#1A488E",
         "black-100": "#100d25",
         "black-200": "#090325",
         "white-100": "#f3f3f3",
